move stats window to tab panel
