+++ src/main/resources/view/StatisticsWindow.fxml
@@ -1,45 +1,29 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <?import javafx.geometry.Insets?>
-<?import javafx.scene.Scene?>
 <?import javafx.scene.control.Label?>
-<?import javafx.scene.image.Image?>
-<?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.StackPane?>
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
-<?import javafx.stage.Stage?>
 
-<fx:root resizable="false" title="Statistics" type="javafx.stage.Stage" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1">
-   <icons>
-      <Image url="@/images/stats-icon.png" />
-   </icons>
-   <scene>
-      <Scene>
-         <AnchorPane prefHeight="400.0" prefWidth="600.0" style="-fx-background-color: #101010;" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1">
-            <children>
-               <VBox prefHeight="200.0" prefWidth="100.0" style="-fx-background-color: #242424;" AnchorPane.bottomAnchor="20.0" AnchorPane.leftAnchor="20.0" AnchorPane.rightAnchor="20.0" AnchorPane.topAnchor="20.0">
-                  <children>
-                     <Label alignment="CENTER" maxWidth="1.7976931348623157E308" text="Top Selling Products" textAlignment="CENTER" textFill="#e797fd" VBox.vgrow="ALWAYS">
-                        <font>
-                           <Font size="17.0" />
-                        </font>
-                     </Label>
-                     <VBox fx:id="productList" minWidth="150.0">
-                        <padding>
-                           <Insets bottom="10" left="10" right="10" top="10" />
-                        </padding>
-                        <children>
-                           <StackPane fx:id="statisticsListPanelPlaceholder" VBox.vgrow="ALWAYS" />
-                        </children>
-                     </VBox>
-                  </children>
-                  <padding>
-                     <Insets top="20.0" />
-                  </padding>
-               </VBox>
-            </children>
-         </AnchorPane>
-      </Scene>
-   </scene>
+
+<fx:root prefHeight="200.0" prefWidth="100.0" style="-fx-background-color: #242424;" type="VBox" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1">
+   <children>
+      <Label alignment="CENTER" maxWidth="1.7976931348623157E308" text="Top Selling Products" textAlignment="CENTER" textFill="#e797fd" VBox.vgrow="ALWAYS">
+         <font>
+            <Font size="17.0" />
+         </font>
+      </Label>
+      <VBox fx:id="productList" minWidth="150.0">
+         <padding>
+            <Insets bottom="10" left="10" right="10" top="10" />
+         </padding>
+         <children>
+            <StackPane fx:id="statisticsListPanelPlaceholder" VBox.vgrow="ALWAYS" />
+         </children>
+      </VBox>
+   </children>
+   <padding>
+      <Insets top="20.0" />
+   </padding>
 </fx:root>
