+++ src/main/resources/view/TransactionListPanel.fxml
@@ -17,9 +17,12 @@
        <TableColumn fx:id="customerCol" prefWidth="125.0" sortable="false" text="Customer" />
        <TableColumn fx:id="productCol" prefWidth="125.0" sortable="false" text="Product" />
          <TableColumn fx:id="dateTimeCol" prefWidth="150.0" sortable="false" text="DateTime" />
-         <TableColumn fx:id="quantityCol" prefWidth="75.0" sortable="false" text="Quantity" />
-         <TableColumn fx:id="amountCol" prefWidth="75.0" sortable="false" text="Amount" />
+         <TableColumn fx:id="quantityCol" sortable="false" text="Quantity" />
+         <TableColumn fx:id="amountCol" sortable="false" text="Amount" />
          <TableColumn fx:id="descriptionCol" prefWidth="300.0" sortable="false" text="Description" />
      </columns>
+      <columnResizePolicy>
+         <TableView fx:constant="CONSTRAINED_RESIZE_POLICY" />
+      </columnResizePolicy>
    </TableView>
 </VBox>
